entrega 0 - Algunos tests y modificaciones chicas en clases

+++ src/main/java/domain/Categoria.java
@@ -4,12 +4,12 @@ public class Categoria {
 
 	private int id;
 	private String nombre;
-	private float cargoFijo;
-	private float cargoVariable;
+	private double cargoFijo;
+	private double cargoVariable;
 	private int limiteInferiorConsumo;
 	private int limiteSuperiorConsumo;
 
-	public Categoria(int id, String nombre, float cargoFijo, float cargoVariable, int limiteInferiorConsumo,
+	public Categoria(int id, String nombre, double cargoFijo, double cargoVariable, int limiteInferiorConsumo,
 			int limiteSuperiorConsumo) {
 		super();
 		this.id = id;
@@ -36,19 +36,19 @@ public class Categoria {
 		this.nombre = nombre;
 	}
 
-	public float getCargoFijo() {
+	public double getCargoFijo() {
 		return cargoFijo;
 	}
 
-	public void setCargoFijo(float cargoFijo) {
+	public void setCargoFijo(double cargoFijo) {
 		this.cargoFijo = cargoFijo;
 	}
 
-	public float getCargoVariable() {
+	public double getCargoVariable() {
 		return cargoVariable;
 	}
 
-	public void setCargoVariable(float cargoVariable) {
+	public void setCargoVariable(double cargoVariable) {
 		this.cargoVariable = cargoVariable;
 	}
